requirements/base.txt
boto3
celery
django
django-celery-beat
django-celery-results
django-cors-headers
django-nested-inline
django-jsoneditor-widget
django-activity-stream
django-jsonfield
django-notifications-hq
django-rest-framework
django-storages
django-thumbnails
python-decouple
sorl-thumbnail
psycopg2
matplotlib
tabulate
redis
requests
PyVimeo
watchdog
wkhtmltopdf